Extract contract ref common code
